backend/requirements.txt
PyMuPDF>=1.23.0
oracledb
python-docx
pandas
numpy
openai
python-dotenv
Pillow
flask
flask_cors
openpyxl

fastapi
uvicorn
sqlalchemy
python-multipart

Flask
requests
google-genai
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
gunicorn
psycopg2-binary
flashtext